requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.6.2=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb9d737c_2
asttokens=2.2.1=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
babel=2.11.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.2=pyha770c72_0
blas=1.0=mkl
bleach=6.0.0=pyhd8ed1ab_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=38.0.4=py39h9ce1e76_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.0.140=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.140=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.0.146=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.10=nompi_h77c792f_102
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.5.1.14=0
giflib=5.2.1=h5eee18b_1
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=hf484d3e_1000
idna=3.4=py39h06a4308_0
igraph=0.10.6=pypi_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_resources=5.10.2=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.5.5=py39hef51801_0
ipython=8.9.0=pyh41d4057_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.4=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=2.11.3=pyhd8ed1ab_2
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hf3d152e_0
jupyter_server=1.23.4=py39h06a4308_0
jupyterlab=3.5.3=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
jupyterlab_widgets=3.0.5=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libclang=10.0.1=default_hb85057a_2
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.5.1.14=0
libcufile-dev=1.5.1.14=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=he513fc3_3
libllvm11=11.1.0=hf817b99_2
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvmlite=0.39.1=py39he621ea3_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=1.1.1=py39h3811e60_3
matplotlib=3.6.2=py39h06a4308_0
matplotlib-base=3.6.2=py39h945d387_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=0.8.4=pyh1a96a4e_1006
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nb_conda_kernels=2.3.1=py39hf3d152e_2
nbclassic=0.5.1=pyhd8ed1ab_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.7.3=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
notebook=6.5.2=pyha770c72_0
notebook-shim=0.2.2=pyhd8ed1ab_0
nsight-compute=2022.4.1.6=0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.56.4=py39h417a72b_0
numpy=1.22.3=py39he7a7128_0
numpy-base=1.22.3=py39hf524024_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.3.0=py39h6a678d5_2
pip=22.3.1=py39h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.6.0=pyhd8ed1ab_0
powerlaw=1.5=pypi_0
prometheus_client=0.16.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=pyha770c72_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.2.0=py39_torch_1.13.0_cu116
pygments=2.14.0=pyhd8ed1ab_0
pynndescent=0.5.8=pyh1a96a4e_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cluster=1.6.0=py39_torch_1.13.0_cu116
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytorch-scatter=2.1.0=py39_torch_1.13.0_cu116
pytorch-sparse=0.6.15=py39_torch_1.13.0_cu116
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py39hb69f2a1_2
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
scikit-learn=1.2.0=py39h6a678d5_0
scipy=1.10.0=py39h14f4228_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.6.3=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.40.1=h5082296_0
stack_data=0.6.2=pyhd8ed1ab_0
tbb=2021.6.0=hdb19cb5_1
terminado=0.17.1=pyh41d4057_0
testpath=0.6.0=pyhd8ed1ab_0
texttable=1.6.7=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=0.13.1=py39_cu116
torchvision=0.14.1=py39_cu116
tornado=6.1=py39hb9d737c_3
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
umap-learn=0.5.3=py39hf3d152e_0
urllib3=1.26.14=py39h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.5.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=pyhd8ed1ab_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.12.1=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
